requirements.txt
aiofiles==24.1.0
aiosqlite==0.21.0
annotated-types==0.7.0
bcrypt==4.3.0
dnspython==2.7.0
email_validator==2.2.0
html5tagger==1.3.0
httptools==0.6.4
idna==3.10
iso8601==2.1.0
multidict==6.1.0
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pypika-tortoise==0.5.0
pytz==2025.1
PyYAML==6.0.2
sanic==24.12.0
Sanic-Cors==2.2.0
sanic-ext==24.12.0
sanic-jwt==1.8.0
sanic-routing==23.12.0
setuptools==76.0.0
tortoise-orm==0.24.2
tracerite==1.1.1
typing_extensions==4.12.2
ujson==5.10.0
uvloop==0.21.0
websockets==15.0.1
